Made some fixes
