+++ android/app/src/main/AndroidManifest.xml
@@ -13,6 +13,8 @@
     <uses-permission android:name="android.permission.READ_CONTACTS" />
     <uses-permission android:name="android.permission.VIBRATE" />
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION"/>
+    <uses-permission android:name="android.permission.WAKE_LOCK" />
+    <uses-permission android:name="android.permission.FOREGROUND_SERVICE" />
 
     <application
         android:name="io.flutter.app.FlutterApplication"
@@ -42,5 +44,4 @@
             android:enabled="true" />
         <service android:name="LocationService" android:enabled="true"/>
     </application>
-    <uses-permission android:name="android.permission.WAKE_LOCK" />
 </manifest>
