+++ android/app/src/main/kotlin/shinobi/code/safely/utility.kt
@@ -5,7 +5,6 @@ import android.util.Log
 import java.lang.Exception
 
 fun sendSms(number: String?, message: String?): Int{
-    Log.e("Inside sendSms", "Okay we're now inside the send sms method, yaaaay")
     return try {
         val smsManager = SmsManager.getDefault()
         smsManager.sendTextMessage(
@@ -15,7 +14,6 @@ fun sendSms(number: String?, message: String?): Int{
             null,
             null
         )
-        Log.e("SendSms:", "Sent sms")
         1
     } catch (ex: Exception) {
         ex.printStackTrace()
