+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: safely
-description: This is my final year project. It is a flutter application for people who might be in danger of being kidnapped&#x2F;raped. This README will be updated as time goes on.
+description: A flutter application for people who might be in danger of being kidnapped&#x2F;raped.
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
